refactor: Consolidate setting/sync UI conventions
