+++ webview/setting/components/Form.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Locale} from '../locales/ja';
 import {Title} from './Title';
-import {Error} from './Error';
+import {Error} from '../../common/components/Error';
 import {URLInput} from './URLInput';
 import {Text} from './Text';
 import {Select} from './Select';
